fix gupnet code
